requirements.txt
Pillow
numpy==1.26.4
black
pre-commit
pyright